requirements.txt
torch>=2.1.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
mlflow>=2.8.0
pyyaml>=6.0
tqdm>=4.66.0
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
